Code Panel solved and Interact Input included
Code Panel had a bug that has been solved. Now, after introducing 4 digits, a delay of 1 seconds happens and it checks if the code is correct or not.
TBD: include sound for fail or success.

Input has been added: Interact. It is used to interact with Interactive Actors. The same Interact action is used to close the text UI for the interactive actors as well as the special UI for those interactive actors which have them.