Highlight Component included
A new Material has been created, called "CB_HighlightMaterial". To make proper use of it, it is necessary to add the "BP_HighlightComponent" to any actor which is desired to be highlighted when being at a certain distance from the character.

In this case, those actors with the Highlight Component will be highlighted with a line around the object.

This has already been implemented in:
- Two cubes in the ClockRoom map.
- Clock in the ClockRoom map (Collisions have been deactivated from the needles to prevent flickering when highlighting)
- Interactive Stars

Another important thing would be that for the Highlight Component to work, the map where the actors are must have a Post Process Volume with the following two parameters set:
- Infinite Bounds (Unbound) checked.
- Post Process Materials -> Add Material -> Select from Asset Reference -> Use "CB_HighlightMaterial".