Changes in Interactive Stars, PlaySoundButton
Now there is a Button called "BP_PlaySoundButton" which has been created to stop the music so that the stars stop blinking, to allow the player guess the puzzle.

"BP_InteractiveStar" has also now the code that makes the stars blink and increase their scale based on the attenuation of the node related to it, using an integer "StarIndex" that goes from 1 to 4.

Both elements need to be configured on the editor, since the PlaySoundButton needs to select the array of BP_InteractiveStars to stop/play. The same way the BP_interactiveStar has a "StarIndex" integer which needs to be placed for those stars that correspond.