Changes on Interactive Star, Clock, Game Mode and Character
Now all the stars shine when the code is wrong or correct (in red or green respectively). This is an event which comes from the Game Mode called "Star Code Check".
The event is called from the character. When the stars are interacted by the player, the stars should shine with the color indicated. Now the "Shine" function receives an input which is the Linear Color in which it is going to shine.

TBD: make the scale change with a Timeline when stars are interacted.